+++ packages/bank-card/src/docs/Component.stories.tsx
@@ -4,15 +4,15 @@ import { text, select } from '@storybook/addon-knobs';
 import {
     getQueryParam,
     stylesStringToObj,
-} from '../../../screenshot-utils/screenshots-story/utils';
+} from '@alfalab/core-components-screenshot-utils/screenshots-story/utils';
 import styles from './preview.module.css';
 
-import { AlfaBankLIcon } from '@alfalab/icons-logotype/AlfaBankLIcon';
+import { AlfaBankSignMIcon } from '@alfalab/icons-logo/AlfaBankSignMIcon';
 import { AlfaInvestEnLogoShortLIcon } from '@alfalab/icons-logotype/AlfaInvestEnLogoShortLIcon';
 import { BankCard } from '@alfalab/core-components-bank-card';
 
 const logo = {
-    default: <AlfaBankLIcon />,
+    default: <AlfaBankSignMIcon width={30} height={40} />,
     custom: <AlfaInvestEnLogoShortLIcon />,
 };