+++ packages/confirmation/src/components/screens/initial/countdown-section.tsx
@@ -1,4 +1,4 @@
-import React, { FC, Fragment,ReactNode, useContext } from 'react';
+import React, { FC, Fragment, ReactNode, useContext } from 'react';
 import cn from 'classnames';
 
 import { Button } from '@alfalab/core-components-button';
@@ -79,7 +79,8 @@ export const CountdownSection: FC<CountdownSectionProps> = ({
 
     return renderText(
         <Fragment>
-            {texts.countdown}{'\u00A0'}
+            {texts.countdown}
+            {'\u00A0'}
             <time className={styles.timeCounter}>{formatMsAsMinutes(timeLeft)}</time>
         </Fragment>,
     );