+++ packages/plate/src/component.screenshots.test.tsx
@@ -3,7 +3,7 @@ import {
     generateTestCases,
     createSpriteStorybookUrl,
     createPreview,
-} from '../../screenshot-utils';
+} from '@alfalab/core-components-screenshot-utils';
 
 const screenshotTesting = setupScreenshotTesting({
     it,