+++ packages/universal-modal/src/components/content/desktop/Component.desktop.tsx
@@ -1,15 +1,28 @@
-import React, { FC } from 'react';
+import React, { FC, useContext } from 'react';
 import cn from 'classnames';
 
+import { UniversalModalContext } from '../../../context/universal-modal-context';
 import { BaseContent, ContentProps } from '../base-content/base-content';
 
-export type ContentDesktopProps = ContentProps & {
+import styles from './desktop.module.css';
+
+export interface ContentDesktopProps extends ContentProps {
     /**
      * Размер (только для desktop версии компонента)
      */
     size?: 's' | 500;
-};
+}
 
-export const ContentDesktop: FC<ContentDesktopProps> = ({ className, ...restProps }) => (
-    <BaseContent className={cn(className)} {...restProps} />
-);
+export const ContentDesktop: FC<ContentDesktopProps> = ({ className, ...restProps }) => {
+    const { hasHeader, hasFooter } = useContext(UniversalModalContext);
+
+    return (
+        <BaseContent
+            className={cn(styles.content, className, {
+                [styles.withHeader]: hasHeader,
+                [styles.withFooter]: hasFooter,
+            })}
+            {...restProps}
+        />
+    );
+};