+++ packages/universal-modal/src/components/content/mobile/Component.mobile.tsx
@@ -1,12 +1,23 @@
-import React, { FC } from 'react';
+import React, { FC, useContext } from 'react';
 import cn from 'classnames';
 
-import { BaseContent, ContentProps } from '../base-content/base-content';
+import { ModalContext } from '../../../Context';
+import { type ContentProps, BaseContent } from '../base-content/base-content';
 
 import styles from './mobile.module.css';
 
 export type ContentMobileProps = ContentProps;
 
-export const ContentMobile: FC<ContentMobileProps> = ({ className, ...restProps }) => (
-    <BaseContent className={cn(styles.content, styles.withFooter, className)} {...restProps} />
-);
+export const ContentMobile: FC<ContentMobileProps> = ({ className, ...restProps }) => {
+    const { hasHeader, hasFooter } = useContext(ModalContext);
+
+    return (
+        <BaseContent
+            className={cn(styles.content, className, {
+                [styles.withHeader]: hasHeader,
+                [styles.withFooter]: hasFooter,
+            })}
+            {...restProps}
+        />
+    );
+};