+++ packages/universal-modal/src/desktop/types/props.ts
@@ -1,8 +1,10 @@
-import { BaseModalProps } from '@alfalab/core-components-base-modal';
+import type { RefObject } from 'react';
 
-import { TMargin } from '../../typings/margin-type';
+import type { BaseModalProps } from '@alfalab/core-components-base-modal';
 
-export type BaseUniversalModalProps = {
+import type { TMargin } from '../../typings/margin-type';
+
+export interface BaseUniversalModalProps {
     /**
      * Расположение по горизонтали и сторона с которой модал “выезжает” при открытии
      * @default center
@@ -26,10 +28,11 @@ export type BaseUniversalModalProps = {
     /**
      * Высота модального окна
      * fullHeight - высота модального окна подстроится под viewport
+     * hugContent - высота модального окна зафиксируется под размер содержимого
      * minHeight - 264
      * @default fullHeight
      */
-    height?: number | 'fullHeight';
+    height?: number | 'fullHeight' | 'hugContent';
 
     /**
      * Наличие оверлея
@@ -41,26 +44,27 @@ export type BaseUniversalModalProps = {
      * Устанавливает отступы модального окна
      */
     margin?: TMargin;
+}
 
+export interface UniversalModalDesktopProps
+    extends BaseUniversalModalProps,
+        Pick<BaseModalProps, 'open'>,
+        Partial<
+            Pick<
+                BaseModalProps,
+                | 'children'
+                | 'dataTestId'
+                | 'className'
+                | 'wrapperClassName'
+                | 'onUnmount'
+                | 'transitionProps'
+                | 'backdropProps'
+                | 'disableBackdropClick'
+                | 'onClose'
+            >
+        > {
     /**
-     * Хэндлер закрытия модалки
+     * Реф контейнера на котором происходит scroll
      */
-    onClose?: () => void;
-};
-
-export type UniversalModalDesktopProps = BaseUniversalModalProps &
-    Pick<
-        BaseModalProps,
-        | 'children'
-        | 'dataTestId'
-        | 'open'
-        | 'className'
-        | 'wrapperClassName'
-        | 'onUnmount'
-        | 'transitionProps'
-        | 'backdropProps'
-    >;
-
-export type ModalBySideProps = UniversalModalDesktopProps;
-
-export type ModalByCenterProps = UniversalModalDesktopProps;
+    scrollableContainerRef?: RefObject<HTMLDivElement>;
+}