+++ packages/universal-modal/src/typings/index.ts
@@ -1,25 +1,27 @@
-import { UniversalModalDesktopProps } from '../desktop/types/props';
-import { UniversalModalMobileProps } from '../mobile/types/props';
+import { UniversalModalDesktopProps } from '../desktop';
+import { UniversalModalMobileProps } from '../mobile';
 
-export type TResponsiveModalContext = {
-    modalWidth?: number;
+export interface UniversalModalContextType {
+    modalWidth: UniversalModalDesktopProps['width'];
     modalHeaderHighlighted?: boolean;
     modalFooterHighlighted?: boolean;
-    setModalWidth?: (width: number) => void;
+    hasHeader: boolean;
+    hasFooter: boolean;
     setModalHeaderHighlighted?: (value: boolean) => void;
     setModalFooterHighlighted?: (value: boolean) => void;
-} | null;
+}
 
-export type UniversalModalResponsiveProps = UniversalModalDesktopProps &
-    UniversalModalMobileProps & {
-        /**
-         * Контрольная точка, с нее начинается desktop версия
-         * @default 1024
-         */
-        breakpoint?: number;
+export interface UniversalModalResponsiveProps
+    extends UniversalModalDesktopProps,
+        UniversalModalMobileProps {
+    /**
+     * Контрольная точка, с нее начинается desktop версия
+     * @default 1024
+     */
+    breakpoint?: number;
 
-        /**
-         * Значение по-умолчанию для хука useMatchMedia
-         */
-        defaultMatchMediaValue?: boolean | (() => boolean);
-    };
+    /**
+     * Значение по-умолчанию для хука useMatchMedia
+     */
+    defaultMatchMediaValue?: boolean | (() => boolean);
+}