+++ packages/vars/src/colors-decorative.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line import/no-webpack-loader-syntax
-import styles from '!!css-loader!@alfalab/core-components/vars/colors-decorative.css';
-
-export default styles.toString();