+++ packages/calendar/CHANGELOG.md
@@ -1,4 +1,134 @@
-# Change Log
+# @alfalab/core-components-calendar
+
+## 8.0.1
+
+### Patch Changes
+
+<sup><time>26.08.2025</time></sup>
+
+#### [#1820](https://github.com/core-ds/core-components/pull/1820)
+
+-   Обновлены `@alfalab/icons-*` пакеты
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-icon-button@7.0.1
+    -   @alfalab/core-components-shared@1.1.0
+    -   @alfalab/core-components-modal@10.0.1
+    -   @alfalab/core-components-button@12.0.1
+    -   @alfalab/core-components-mq@5.0.1
+    -   @alfalab/core-components-typography@5.0.1
+
+## 8.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-button@12.0.0
+    -   @alfalab/core-components-icon-button@7.0.0
+    -   @alfalab/core-components-modal@10.0.0
+    -   @alfalab/core-components-mq@5.0.0
+    -   @alfalab/core-components-shared@1.0.0
+    -   @alfalab/core-components-typography@5.0.0
+
+## 7.20.0
+
+### Minor Changes
+
+<sup><time>04.08.2025</time></sup>
+
+### [#1788](https://github.com/core-ds/core-components/pull/1788)
+
+-   Для компонента 'CalendarDesktop' с видом (selectorView='month-only') добавлен параметр 'hideDisabledArrows', позволяющий управлять видимостью кнопок.
+
+## 7.19.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   modal@9.9.20
+
+## 7.19.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   modal@9.9.19
+
+## 7.19.0
+
+### Minor Changes
+
+<sup><time>23.06.2025</time></sup>
+
+### [#1725](https://github.com/core-ds/core-components/pull/1725)
+
+-   Добавлена возможность управлять кнопкой назад для CalendarMobile (пропсы hasBackButton, onBack)
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+    -   button@11.11.10
+    -   modal@9.9.18
+    -   typography@4.14.3
+    -   icon-button@6.11.13
+
+## 7.18.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   modal@9.9.17
+
+## 7.18.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+    -   button@11.11.9
+    -   modal@9.9.16
+    -   typography@4.14.2
+    -   icon-button@6.11.12
+
+## 7.18.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.1
+    -   modal@9.9.15
+
+## 7.18.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.0
+    -   modal@9.9.14
+
+## 7.18.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.0
+    -   modal@9.9.13
+    -   button@11.11.8
+    -   typography@4.13.1
+    -   icon-button@6.11.11
 
 ## 7.18.2