+++ packages/gallery/CHANGELOG.md
@@ -1,4 +1,153 @@
-# Change Log
+# @alfalab/core-components-gallery
+
+## 6.0.2
+
+### Patch Changes
+
+<sup><time>26.08.2025</time></sup>
+
+#### [#1820](https://github.com/core-ds/core-components/pull/1820)
+
+-   Обновлены `@alfalab/icons-*` пакеты
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-icon-button@7.0.1
+    -   @alfalab/core-components-tooltip@8.0.1
+    -   @alfalab/core-components-base-modal@6.0.1
+    -   @alfalab/core-components-button@12.0.1
+    -   @alfalab/core-components-mq@5.0.1
+    -   @alfalab/core-components-spinner@5.0.1
+    -   @alfalab/core-components-typography@5.0.1
+
+## 6.0.1
+
+### Patch Changes
+
+<sup><time>15.08.2025</time></sup>
+
+#### [#1794](https://github.com/core-ds/core-components/pull/1794)
+
+-   Добавлен fallback шаринга по ссылке для изображений из других источников
+
+## 6.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-base-modal@6.0.0
+    -   @alfalab/core-components-button@12.0.0
+    -   @alfalab/core-components-icon-button@7.0.0
+    -   @alfalab/core-components-icon-view@4.0.0
+    -   @alfalab/core-components-mq@5.0.0
+    -   @alfalab/core-components-spinner@5.0.0
+    -   @alfalab/core-components-tooltip@8.0.0
+    -   @alfalab/core-components-typography@5.0.0
+
+## 5.12.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.10.0
+    -   icon-view@3.11.0
+
+## 5.12.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   tooltip@7.9.17
+
+## 5.12.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-view@3.10.4
+
+## 5.12.8
+
+### Patch Changes
+
+<sup><time>04.07.2025</time></sup>
+
+### [#1756](https://github.com/core-ds/core-components/pull/1756)
+
+-   safe-area отступы перенесены из modal в container
+
+## 5.12.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   icon-view@3.10.3
+    -   base-modal@5.9.4
+    -   button@11.11.10
+    -   spinner@4.0.9
+    -   tooltip@7.9.16
+    -   typography@4.14.3
+    -   icon-button@6.11.13
+
+## 5.12.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   base-modal@5.9.3
+    -   icon-view@3.10.2
+    -   tooltip@7.9.15
+
+## 5.12.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   base-modal@5.9.2
+    -   button@11.11.9
+    -   spinner@4.0.8
+    -   tooltip@7.9.14
+    -   typography@4.14.2
+    -   icon-button@6.11.12
+
+## 5.12.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.1
+    -   tooltip@7.9.13
+
+## 5.12.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.0
+    -   tooltip@7.9.12
+
+## 5.12.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   base-modal@5.9.1
+    -   button@11.11.8
+    -   spinner@4.0.7
+    -   tooltip@7.9.11
+    -   typography@4.13.1
+    -   icon-button@6.11.11
 
 ## 5.12.1