+++ packages/mq/CHANGELOG.md
@@ -1,4 +1,39 @@
-# Change Log
+# @alfalab/core-components-mq
+
+## 5.0.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-shared@1.1.0
+
+## 5.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Пакет `@alfalab/core-config` перенесен в `@alfalab/core-components-config`. При обновлении необходимо заменить `as-is`.
+
+Тип `CoreConfigContext` переименован в `CoreConfigContextValue`.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-config@1.0.0
+    -   @alfalab/core-components-shared@1.0.0
 
 ## 4.4.1