+++ packages/steps/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @alfalab/core-components-steps
 
+## 2.0.1
+
+### Patch Changes
+
+<sup><time>26.08.2025</time></sup>
+
+#### [#1820](https://github.com/core-ds/core-components/pull/1820)
+
+-   Обновлены `@alfalab/icons-*` пакеты
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-shared@1.1.0
+
+## 2.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-badge@6.0.0
+    -   @alfalab/core-components-shared@1.0.0
+
+## 1.8.4
+
+### Patch Changes
+
+<sup><time>04.07.2025</time></sup>
+
+### [#1755](https://github.com/core-ds/core-components/pull/1755)
+
+-   Удален отступ у последнего дочернего элемента списка Steps
+
+## 1.8.3
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+
+## 1.8.2
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+
+## 1.8.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.0
+
 ## 1.8.0
 
 ### Minor Changes